IOW-307 Incremental fixes to DV legend pieces, moved common markers.js piece to d3-rendering dir.
